Update README.md
It's a Jarvius Virtual Assistant where only Backend is available. As i have the basic information of Python. 

This Assistant works similar to alexa. You can operate it in your system (PC, Laptop, etc) and can ask questions from it.
As it is created with basic information of Python it just have few instructions